requirements.txt
pandas==2.2.3
mysql-connector-python==9.3.0
psycopg2-binary==2.9.10
loguru==0.7.3